requirements.txt
python-telegram-bot>=20.0,<22.0
requests